Add support for Ui class and Parser class
