+++ data/sai.txt
@@ -1,2 +1,2 @@
 T | 0 | test
-E | 0 | abc | 2020-12-02T18:00 | 2020-12-03T19:00
+T | 0 | name
